+++ pybamm/parameters/lead_acid_parameters.py
@@ -20,7 +20,6 @@ class LeadAcidParameters(BaseParameters):
     """
 
     def __init__(self):
-
         # Get geometric, electrical and thermal parameters
         self.geo = pybamm.geometric_parameters
         self.elec = pybamm.electrical_parameters
@@ -67,6 +66,7 @@ class LeadAcidParameters(BaseParameters):
         # Electrical
         self.I_typ = self.elec.I_typ
         self.Q = self.elec.Q
+        self.R_contact = self.elec.R_contact
         self.C_rate = self.elec.C_rate
         self.n_electrodes_parallel = self.elec.n_electrodes_parallel
         self.n_cells = self.elec.n_cells